% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Soft_Hybrid_Imputation.R
\name{run_soft_hybrid_all_groups}
\alias{run_soft_hybrid_all_groups}
\title{Run Soft Hybrid Across Groups}
\usage{
run_soft_hybrid_all_groups(
  raw_list,
  rf_list,
  minProb_list,
  lambda = 0.5,
  a = 10,
  b = 5,
  visualize = TRUE
)
}
\arguments{
\item{raw_list}{A named list of raw matrices.}

\item{rf_list}{A named list of lists with element \code{imputed_data}.}

\item{minProb_list}{A named list of lists with element \code{imputed_data}.}

\item{lambda}{Trade-off between missing-rate and intensity (0–1, default 0.5).}

\item{a}{Steepness for missing-rate sigmoid (default 10).}

\item{b}{Steepness for intensity sigmoid (default 5).}

\item{visualize}{Logical. If TRUE, saves a TIFF + Rdata of the weight plot.}
}
\value{
A named list of matrices with soft-hybrid imputations.
}
\description{
Run Soft Hybrid Across Groups
}
